fix bug: dsn==0

+++ mptcpproxy_util.h
@@ -446,7 +446,7 @@ struct session;
 
 struct subflow{
 	//new+++++++++++++++
-	uint16_t highest_org_sn_loc;//used for retrx
+	uint32_t highest_org_sn_loc;//used for retrx
 	//new---------------
 
 	struct fourtuple ft;//key
